requirements.repo.txt
--extra-index-url https://test.pypi.org/simple/

ynov-python-2023-2024-data-processing
ynov-python-2023-2024-data-analysis
ynov-python-2023-2024-sales-prediction
